refactor(research): extract completion counts into own module
research.tsx exceeded the 300-line limit after the skipResponseSearch
addition (prettier re-expands condensed lines, so inline condensation is
not durable). Extract the completion-counts feature into its own module,
paired with the existing research-counts test file.
